requirements.txt
Flask
gunicorn
openai
python-dotenv
Flask-Cors
